refactor build and publish mix tasks
